fixes #878
